Flytter ut data-generator og legger til allowedStroage
